+++ src/main/java/mate/academy/dto/request/GoodsRequestDto.java
@@ -1,5 +1,6 @@
 package mate.academy.dto.request;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,5 +8,5 @@ import lombok.Setter;
 @Setter
 public class GoodsRequestDto {
     private String name;
-    private Double cost;
+    private BigDecimal cost;
 }
